requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-community
langchain-openai
streamlit_option_menu

-e .